+++ mex/extractors/seq_repo/main.py
@@ -1,6 +1,6 @@
 from mex.common.cli import entrypoint
 from mex.common.ldap.extract import get_merged_ids_by_query_string
-from mex.common.ldap.models.person import LDAPPersonWithQuery
+from mex.common.ldap.models import LDAPPersonWithQuery
 from mex.common.ldap.transform import transform_ldap_persons_with_query_to_mex_persons
 from mex.common.models import (
     AccessPlatformMapping,
@@ -54,9 +54,12 @@ def seq_repo_source() -> list[SeqRepoSource]:
 @asset(group_name="seq_repo")
 def seq_repo_latest_source(
     seq_repo_source: list[SeqRepoSource],
+    extracted_primary_source_seq_repo: ExtractedPrimarySource,
 ) -> dict[str, SeqRepoSource]:
     """Filter latest sources from seq-repo source."""
-    return filter_sources_on_latest_sequencing_date(seq_repo_source)
+    return filter_sources_on_latest_sequencing_date(
+        seq_repo_source, extracted_primary_source_seq_repo
+    )
 
 
 @asset(group_name="seq_repo")