+++ tests/ff_projects/test_filter.py
@@ -1,7 +1,10 @@
 from mex.common.models import ExtractedPrimarySource
 from mex.common.types import MergedOrganizationalUnitIdentifier
 from mex.extractors.ff_projects.extract import extract_ff_projects_sources
-from mex.extractors.ff_projects.filter import filter_and_log_ff_projects_sources
+from mex.extractors.ff_projects.filter import (
+    filter_and_log_ff_projects_sources,
+    filter_out_duplicate_source_ids,
+)
 
 
 def test_filter_and_log_ff_projects_sources(
@@ -42,3 +45,17 @@ def test_filter_and_log_ff_projects_sources(
         "Skipped Laufzeit von before skipped years",
         "Skipped Laufzeit bis before skipped years",
     ]
+
+
+def test_filter_out_duplicate_source_ids() -> None:
+    ff_proj_liste_sources = list(extract_ff_projects_sources())
+
+    assert len(ff_proj_liste_sources) == 21
+    unfiltered_ids = [s.lfd_nr for s in ff_proj_liste_sources]
+    assert unfiltered_ids.count("20") == 2
+
+    filtered = filter_out_duplicate_source_ids(ff_proj_liste_sources)
+
+    assert len(filtered) == 19
+    filtered_ids = [s.lfd_nr for s in filtered]
+    assert filtered_ids.count("20") == 0